generator now produces data for cid-related benchmarking
