--- junk/cidify/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "cidify"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[dependencies]
-aquavm-air = { path = "../../air", features = ["gen_signatures", "check_signatures"] }
-air-interpreter-cid = { path = "../../crates/air-lib/interpreter-cid" }
-air-interpreter-data = { path = "../../crates/air-lib/interpreter-data" }
-air-interpreter-signatures = { path = "../../crates/air-lib/interpreter-signatures" }
-air-test-utils = { path = "../../crates/air-lib/test-utils" }
-clap = { version = "4.1.13", features = ["derive"] }
-serde = { version = "1.0.158", features = ["derive"] }
-serde_json = "1.0.94"
