+++ junk/gen-bench-data/gen_benchmark_data.sh
@@ -7,18 +7,20 @@ set -e
 echo "Pre-build a binary..." >&2
 cargo build --quiet --release
 
-for bench in canon-map-key-by-lens \
+for bench in big-values-data \
+            canon-map-key-by-lens \
             canon-map-key-element-by-lens \
             canon-map-multiple-keys \
             canon-map-single-key \
             canon-map-scalar-multiple-keys \
             canon-map-scalar-single-key \
-            populate-map-multiple-keys \
-            populate-map-single-key \
+            dashboard network-explore \
+            long-data \
             multiple-cids10 \
             multiple-peers8 \
             multiple-sigs30 \
-            dashboard network-explore;
+            populate-map-multiple-keys \
+            populate-map-single-key ;
 do
     echo "Generating ${bench} ..." >&2
     DESTDIR="../../benches/performance_metering/${bench}/"
