+++ junk/gen-bench-data/src/cid_benchmarking/anomaly_long.json
